Conflicts:
	.travis.yml
	build.gradle
	softwareconfiguration/mlplan/weka/src/example/java/ai/libs/mlplan/examples/multiclass/weka/MLPlanGraphVisualizationExample.java

+++ JAICore/jaicore-search/src/main/java/ai/libs/jaicore/search/algorithms/mdp/mcts/spuct/SPUCT.java
@@ -6,7 +6,7 @@ import ai.libs.jaicore.search.probleminputs.IMDP;
 
 public class SPUCT<N, A> extends MCTS<N, A> {
 
-	public SPUCT(final IMDP<N, A, Double> input, final IPolicy<N, A> defaultPolicy, final double bigD, final double maxIterations, final double gamma, final double epsilon, final boolean tabooExhaustedNodes) {
+	public SPUCT(final IMDP<N, A, Double> input, final IPolicy<N, A> defaultPolicy, final double bigD, final int maxIterations, final double gamma, final double epsilon, final boolean tabooExhaustedNodes) {
 		super(input, new SPUCBPolicy<>(gamma, bigD), defaultPolicy, maxIterations, gamma, epsilon, tabooExhaustedNodes);
 	}
 }
